login and register components uit App.js halen
4 components uit App.js gehaald en in Auth.js gezet
